Refactor display and touch handling, update configuration files, and enhance emulator logging

+++ lib/core/include/cyd_display.h
@@ -5,13 +5,13 @@
 
 class CYDDisplay : public IDisplay {
 public:
-    explicit CYDDisplay(uint8_t backlight_pin, uint8_t pwm_channel = 0);
+    explicit CYDDisplay(int backlight_pin, uint8_t pwm_channel = 0);
 
     bool begin() override;
     void setBacklightPercent(uint8_t percent) override;
 
 private:
-    uint8_t backlight_pin_;
+    int     backlight_pin_;
     uint8_t pwm_channel_;
 
     uint32_t percentToDuty(uint8_t percent) const;
